Remove PretrainTransformer Vectorizer to avoid Pytest Error (#179)
