+++ pecos/utils/featurization/text/vectorizers.py
@@ -11,40 +11,13 @@
 import json
 import logging
 import os
-import pathlib
 import pickle
 from abc import ABCMeta
 
 import numpy as np
-import torch
 from pecos.core import clib
-from pecos.utils import torch_util
 from sklearn.feature_extraction.text import HashingVectorizer, TfidfVectorizer
-from sklearn.preprocessing import normalize
-from torch.utils.data import DataLoader, SequentialSampler, TensorDataset
-from transformers import (
-    AlbertConfig,
-    AlbertModel,
-    AlbertTokenizer,
-    BertConfig,
-    BertModel,
-    BertTokenizer,
-    DistilBertConfig,
-    DistilBertModel,
-    DistilBertTokenizer,
-    RobertaConfig,
-    RobertaModel,
-    RobertaTokenizer,
-    XLMRobertaConfig,
-    XLMRobertaModel,
-    XLMRobertaTokenizer,
-    XLMConfig,
-    XLMModel,
-    XLMTokenizer,
-    XLNetConfig,
-    XLNetModel,
-    XLNetTokenizer,
-)
+
 
 vectorizer_dict = {}
 
@@ -514,329 +487,3 @@ class SklearnHashing(Vectorizer):
         # Indices must be sorted for C++ batch code to work
         result.sort_indices()
         return result
-
-
-class TransformerModelClass(object):
-    """Utility class for representing a Transformer and tokenizer."""
-
-    def __init__(self, config_class, model_class, tokenizer_class):
-        """Initialization
-
-        Args:
-            config_class (transformers.configuration_utils.PretrainedConfig)
-            model_class (transformers.modeling_utils.PreTrainedModel)
-            tokenizer_class (transformers.tokenization_utils.PreTrainedTokenizer)
-        """
-        self.config_class = config_class
-        self.model_class = model_class
-        self.tokenizer_class = tokenizer_class
-
-
-class PretrainedTransformer(Vectorizer):
-    """Vectorizer with a variety of Transformer models."""
-
-    MODEL_CLASSES = {
-        "albert": TransformerModelClass(AlbertConfig, AlbertModel, AlbertTokenizer),
-        "bert": TransformerModelClass(BertConfig, BertModel, BertTokenizer),
-        "distilbert": TransformerModelClass(
-            DistilBertConfig,
-            DistilBertModel,
-            DistilBertTokenizer,
-        ),
-        "roberta": TransformerModelClass(RobertaConfig, RobertaModel, RobertaTokenizer),
-        "xlm-roberta": TransformerModelClass(
-            XLMRobertaConfig, XLMRobertaModel, XLMRobertaTokenizer
-        ),
-        "xlm": TransformerModelClass(XLMConfig, XLMModel, XLMTokenizer),
-        "xlnet": TransformerModelClass(XLNetConfig, XLNetModel, XLNetTokenizer),
-    }
-
-    def __init__(self, model=None, tokenizer=None, transformer_options=None):
-        """Initialization
-
-        Args:
-            model (transformers.modeling_utils.PreTrainedModel)
-            tokenizer (transformers.tokenization_utils.PreTrainedTokenizer)
-            transformer_options (dict)
-        """
-
-        self.tokenizer = tokenizer
-        self.model = model
-        self.transformer_options = transformer_options
-
-    def save(self, vectorizer_folder):
-        """Save the pre-trained model and tokenizer to disk.
-
-        Args:
-            vectorizer_folder (str): Folder to save to.
-        """
-
-        os.makedirs(vectorizer_folder, exist_ok=True)
-        model_folder = os.path.join(vectorizer_folder, "model")
-        tokenizer_folder = os.path.join(vectorizer_folder, "tokenizer")
-        os.makedirs(model_folder, exist_ok=True)
-        os.makedirs(tokenizer_folder, exist_ok=True)
-        # this creates transformer_options.json
-        with open(
-            os.path.join(vectorizer_folder, "transformer_options.json"),
-            "w",
-            encoding="utf-8",
-        ) as fout:
-            fout.write(json.dumps(self.transformer_options))
-        # this creates config.json, pytorch_model.bin
-        self.model.save_pretrained(model_folder)
-        # this creates added_tokens.json, special_tokens_map.json, tokenizer_config.json, vocab.txt
-        self.tokenizer.save_pretrained(tokenizer_folder)
-
-    @classmethod
-    def load(cls, vectorizer_folder):
-        """Load the pre-trained model and tokenizer from disk.
-
-        Args:
-            vectorizer_folder (str): Folder which was saved to using `PretrainedTransformer.save`.
-
-        Returns:
-            PretrainedTransformer: Loaded object.
-        """
-
-        model_folder = os.path.join(vectorizer_folder, "model")
-        tokenizer_folder = os.path.join(vectorizer_folder, "tokenizer")
-
-        assert os.path.isdir(model_folder), "pre-trained model folder {} does not exist".format(
-            model_folder
-        )
-        assert os.path.isdir(
-            tokenizer_folder
-        ), "pre-trained tokenizer folder {} does not exist".format(tokenizer_folder)
-
-        # Load from transformer_options.json
-        with open(
-            os.path.join(vectorizer_folder, "transformer_options.json"),
-            "r",
-            encoding="utf-8",
-        ) as fin:
-            transformer_options = json.loads(fin.read())
-        dnn_type = cls.MODEL_CLASSES[transformer_options["transformer_type"]]
-        # Load from config.json
-        model_config = dnn_type.config_class.from_pretrained(model_folder)
-        # Load from vocab.txt [,added_tokens.json, special_tokens_map.json, tokenizer_config.json]
-        tokenizer = dnn_type.tokenizer_class.from_pretrained(tokenizer_folder)
-        # Load from pytorch_model.bin
-        model = dnn_type.model_class.from_pretrained(model_folder, config=model_config)
-        return cls(model, tokenizer, transformer_options)
-
-    @classmethod
-    def train(cls, trn_corpus, config=None, dtype=np.float32):
-        """Download a Transformer model.
-
-        Args:
-            trn_corpus (list): Training corpus in the form of a list of strings.
-            config (dict): Dictionary containing keywords/value for training. The keywords are:
-                transformer_type (str): Transformer model type (default 'bert').
-
-                norm (str): embedding normalization method, set to None to skip {'l1','l2'} (default 'l2')
-
-                pooling (str): pooling method {'mean','mask-mean','first','last','cls'} (default 'mean')
-
-                model_shortcut: can be either
-                        [1] str (default): pre-trained transformer model name shortcut (default 'bert-base-uncased')
-                                model configuration & tokenizer & model weights will be downloaded or load from
-                                cache_dir (if previously downloaded).
-                        [2] dictionary: a dictionary containing paths to load model configuration, vocabulary
-                                for tokenizer and pre-trained model weights
-                                (e.g. {"config_path": "path/to/config.json",
-                                        "vocab_path": "path/to/vocab.txt",
-                                        "weight_path": "path/to/pytorch_model.bin" })
-
-                do_fine_tune (bool): whether to fine tune the pre-trained Transformer (default False). Not yet implemented.
-
-                cache_dir (str): cache directory to save downloaded model, set to None (default) to use a temporary folder.
-
-        Returns:
-            PretrainedTransformer: Pre-trained model.
-        """
-
-        defaults = {
-            "transformer_type": "bert",
-            "norm": "l2",
-            "pooling": "mean",
-            "model_shortcut": "bert-base-uncased",
-            "do_fine_tune": False,
-            "cache_dir": None,
-        }
-        config = {**defaults, **config}
-        transformer_type = config["transformer_type"]
-        model_shortcut = config["model_shortcut"]
-        cache_dir = config["cache_dir"]
-
-        assert (
-            transformer_type in cls.MODEL_CLASSES
-        ), "Unsupported model type [{}], available Transformer model types: {}".format(
-            transformer_type, cls.MODEL_CLASSES.keys()
-        )
-        dnn_type = cls.MODEL_CLASSES[transformer_type]
-
-        if isinstance(model_shortcut, str):  # download everything
-            # download model config
-            model_config = dnn_type.config_class.from_pretrained(
-                model_shortcut, cache_dir=cache_dir
-            )
-            # download tokenizer
-            tokenizer = dnn_type.tokenizer_class.from_pretrained(
-                model_shortcut, cache_dir=cache_dir
-            )
-            # download model weights
-            model = dnn_type.model_class.from_pretrained(
-                model_shortcut, config=model_config, cache_dir=cache_dir
-            )
-        else:  # load everything locally
-            model_config_path = model_shortcut["config_path"]
-            vocab_path = model_shortcut["vocab_path"]
-            weight_path = model_shortcut["weight_path"]
-            assert pathlib.Path(model_config_path).is_file(), "model config not found at {}".format(
-                model_config_path
-            )
-            model_config = dnn_type.config_class.from_pretrained(model_config_path)
-            assert pathlib.Path(vocab_path).is_file(), "vocabulary not found at {}".format(
-                vocab_path
-            )
-            tokenizer = dnn_type.tokenizer_class.from_pretrained(vocab_path)
-            assert pathlib.Path(weight_path).is_file(), "model weights not found at {}".format(
-                weight_path
-            )
-            model = dnn_type.model_class.from_pretrained(weight_path, config=model_config)
-
-        if config["do_fine_tune"]:
-            # train the model!
-            raise NotImplementedError("Training for Transformer vectorizer not implemented!")
-
-        return cls(model, tokenizer, config)
-
-    def predict(
-        self,
-        corpus,
-        batch_size=8,
-        truncate_length=300,
-        use_gpu_if_available=True,
-        **kwargs,
-    ):
-        """Vectorizer a corpus.
-
-        Args:
-            corpus (list): List of strings to vectorize.
-            batch_size (int, optional): Default is 8.
-            truncate_length (int, optional): Default is 300.
-            use_gpu_if_available (bool, optional): Default is True.
-
-        Returns:
-            numpy.ndarray: Matrix of features.
-        """
-
-        if self.model.config.max_position_embeddings > 0:
-            truncate_length = min(truncate_length, self.model.config.max_position_embeddings)
-
-        # generate feature batches
-        feature_tensors = self.tokenizer.batch_encode_plus(
-            batch_text_or_text_pairs=corpus,
-            return_tensors="pt",
-            return_attention_mask=True,
-            return_token_type_ids=True,
-            add_special_tokens=True,
-            max_length=truncate_length,
-            truncation=True,
-            padding="longest",
-        )
-        # setup device
-        device, n_active_gpu = torch_util.setup_device(use_gpu_if_available=use_gpu_if_available)
-        # start eval
-        transformer_type = self.transformer_options["transformer_type"]
-        norm = self.transformer_options["norm"]
-        pooling = self.transformer_options["pooling"]
-
-        batch_size = batch_size * max(1, n_active_gpu)
-        data = TensorDataset(
-            feature_tensors["input_ids"],
-            feature_tensors["attention_mask"],
-            feature_tensors["token_type_ids"],
-        )
-
-        sampler = SequentialSampler(data)
-        dataloader = DataLoader(data, sampler=sampler, batch_size=batch_size, num_workers=4)
-
-        # multi-gpu eval
-        if n_active_gpu > 1 and not isinstance(self.model, torch.nn.parallel.DataParallel):
-            model = torch.nn.parallel.DataParallel(self.model)
-        else:
-            model = self.model
-
-        model.eval()
-        model.to(device)
-        embeddings = []
-        for batch in dataloader:
-            batch = tuple(t.to(device) for t in batch)
-
-            with torch.no_grad():
-                inputs = {
-                    "input_ids": batch[0],
-                    "attention_mask": batch[1],
-                    "token_type_ids": batch[2],
-                }
-                if transformer_type == "distilbert":
-                    outputs = model(
-                        input_ids=inputs["input_ids"],
-                        attention_mask=inputs["attention_mask"],
-                    )
-                elif transformer_type in [
-                    "bert",
-                    "roberta",
-                    "xlm-roberta",
-                    "albert",
-                    "xlm",
-                    "xlnet",
-                ]:
-                    outputs = model(
-                        input_ids=inputs["input_ids"],
-                        attention_mask=inputs["attention_mask"],
-                        token_type_ids=inputs["token_type_ids"],
-                    )
-                else:
-                    raise NotImplementedError(
-                        "Unsupported transformer_type {}".format(transformer_type)
-                    )
-
-                # get the embeddings from model output
-                # REF: https://huggingface.co/transformers/v2.3.0/model_doc/bert.html#bertmodel
-                # For bert,roberta,xlm-roberta,albert:  outputs = last_hidden_states, pooled_output, (hidden_states)
-                # For xlm,xlnet,distilbert: outputs = last_hidden_states, (hidden_states), (attentions)
-                if pooling == "mean":
-                    pooled_output = outputs[0].mean(dim=1)
-                elif pooling == "mask-mean":
-                    last_hidden_states = torch_util.apply_mask(outputs[0], inputs["attention_mask"])
-                    pooled_output = last_hidden_states.sum(dim=1)
-                    masked_length = inputs["attention_mask"].sum(dim=1)
-                    pooled_output = pooled_output / masked_length.unsqueeze(1).float()
-                elif pooling == "first":
-                    pooled_output = outputs[0][:, 0, :]
-                elif pooling == "last":
-                    pooled_output = outputs[0][:, -1, :]
-                elif pooling == "cls":
-                    assert transformer_type in [
-                        "bert",
-                        "roberta",
-                        "xlm-roberta",
-                        "albert",
-                    ], "Only {} models have [CLS] token.".format(
-                        ["bert", "roberta", "xlm-roberta", "albert"]
-                    )
-                    # get the [CLS] embedding for the document
-                    pooled_output = outputs[1]
-                else:
-                    raise NotImplementedError("Unsupported pooling method {}".format(pooling))
-
-                embeddings.append(pooled_output.cpu().numpy())
-
-        # construct dense output
-        embeddings = np.concatenate(embeddings, axis=0)
-        if norm is not None:
-            embeddings = normalize(embeddings, norm=norm, axis=1, copy=False)
-        return embeddings
