--- test/tst-data/utils/data/dummy_transformer_config.json
@@ -1,36 +1,0 @@
-{
-  "architectures": [
-    "BertForMaskedLM"
-  ],
-  "attention_probs_dropout_prob": 0.1,
-  "finetuning_task": null,
-  "hidden_act": "gelu",
-  "hidden_dropout_prob": 0.1,
-  "hidden_size": 8,
-  "id2label": {
-    "0": "LABEL_0",
-    "1": "LABEL_1"
-  },
-  "initializer_range": 0.02,
-  "intermediate_size": 10,
-  "is_decoder": false,
-  "label2id": {
-    "LABEL_0": 0,
-    "LABEL_1": 1
-  },
-  "layer_norm_eps": 1e-12,
-  "max_position_embeddings": 256,
-  "model_type": "bert",
-  "num_attention_heads": 1,
-  "num_hidden_layers": 1,
-  "num_labels": 2,
-  "output_attentions": false,
-  "output_hidden_states": true,
-  "output_past": true,
-  "pad_token_id": 0,
-  "pruned_heads": {},
-  "torchscript": false,
-  "type_vocab_size": 2,
-  "use_bfloat16": false,
-  "vocab_size": 15
-}
