--- test/tst-data/utils/data/transformer_vectorizer_config.json
@@ -1,10 +1,0 @@
-{
-    "type": "pretrainedtransformer",
-    "kwargs": {
-        "model_shortcut": {
-            "config_path": "test/tst-data/utils/data/dummy_transformer_config.json",
-            "vocab_path": "test/tst-data/utils/data/vocab.txt",
-            "weight_path": "test/tst-data/utils/data/dummy_transformer_model.bin"
-        }
-    }
-}
