--- test/tst-data/utils/data/vocab.txt
@@ -1,15 +1,0 @@
-[PAD]
-[UNK]
-[CLS]
-[SEP]
-[MASK]
-red
-apple
-iphone
-11
-pro
-case
-samsung
-galaxy
-s10
-plus
